feat: publish both feeds
