+++ _lambda/deploy-succeeded.js
@@ -1,107 +1,17 @@
 const fetch = require('node-fetch');
-const Twitter = require('twitter');
-const Entities = require('html-entities').AllHtmlEntities;
+const { URL } = process.env;
 
-require('dotenv').config;
+const POSTS = `${URL}/.netlify/functions/publish-posts`;
+const NOTES = `${URL}/.netlify/functions/publish-notes`;
 
-const FEED = 'https://danmatthew.co.uk/shorts/feed.json';
-
-const twitter = new Twitter({
-  consumer_key: process.env.TWITTER_CONSUMER_KEY,
-  consumer_secret: process.env.TWITTER_CONSUMER_SECRET,
-  access_token_key: process.env.TWITTER_ACCESS_TOKEN_KEY,
-  access_token_secret: process.env.TWITTER_ACCESS_TOKEN_SECRET,
-});
-
-const handleError = (error) => {
-  console.error(error);
-  const msg = Array.isArray(error) ? error[0].message : error.message;
+exports.handler = async (event, context) => {
+  const triggerPosts = await fetch(POSTS);
+  const triggerNotes = await fetch(NOTES);
 
   return {
-    statusCode: 422,
-    body: String(msg),
+    statusCode: 200,
+    body: JSON.stringify({
+      data: context,
+    }),
   };
-};
-
-const status = (code, message) => {
-  console.log(message);
-
-  return {
-    statusCode: code,
-    body: message,
-  };
-};
-
-const processNotes = async (notes) => {
-  const postItems = notes.items;
-
-  if (!postItems.length) {
-    return status(404, 'No notes to process');
-  }
-
-  const latestNote = postItems[0];
-
-  if (!latestNote.syndicate) {
-    return status(400, 'Latest note has disabled syndication. No action taken');
-  }
-
-  try {
-    const query = await twitter.get('search/tweets', { q: latestNote.url });
-
-    if (query.statuses && query.statuses.length === 0) {
-      return publishNote(latestNote);
-    } else {
-      return status(400, 'Latest note was already syndicated. No action taken');
-    }
-  } catch (error) {
-    return handleError(error);
-  }
-};
-
-const prepareStatusText = (note) => {
-  const maxLength = 280 - 3 - 1 - 23 - 20;
-  const entities = new Entities();
-
-  let text = note.content_html.trim().replace(/<[^>]+>/g, '');
-  text = entities.decode(text);
-
-  if (text.length > maxLength) {
-    text = text.substring(0, maxLength) + '…';
-  }
-
-  text += ' ' + note.url;
-
-  if (note.link && note.link.length) {
-    text += ' ' + note.link;
-  }
-
-  return text;
-};
-
-const publishNote = async (note) => {
-  try {
-    const statusText = prepareStatusText(note);
-
-    const tweet = await twitter.post('statuses/update', {
-      status: statusText,
-    });
-
-    if (tweet) {
-      return status(
-        200,
-        `Note ${note.date_published} successfully posted to Twitter`
-      );
-    } else {
-      return status(422, 'Error posting to Twitter API');
-    }
-  } catch (error) {
-    return handleError(error);
-  }
-};
-
-exports.handler = async () => {
-  return fetch(FEED)
-    .then((response) => response.json())
-    .then(processNotes)
-    .catch(handleError);
 };
